Django Challenge Submission Assignment
Created a project in Pycharm named "DjangoUniversity in a Virtualenv virtual environment.  Utilized Django template files to set up a database to migrate imputed data to. Created the class "djangoClasses within the model.py with four attributes that applied the imputed data to the set table with values in each column named by the four attributes Title, Course Number, Instructor Name and Duration.  Registered the models.py model in the admin.py to register the file with the project.  Created three objects from the djangoClasses class and assign their attribute values.  Added the classApp name to the Main Project setting.py file.  Made a final migration and migrate for the three objects created.  Can view and modify objects in the admin panel in a browser window after runserver is ran from the Command Prompt.