Django Challenge Submission Assignment
Created a new project in PyCharm named "DjangoUniversity" in a Virtualenv virtual environment.  Install django to the virtual environment.  With the PyCharm terminal set up the Django template files to start project Classes.  Created an app within the project call classApp.  Within the model.py made a class called "djangoClasses" with four attributes: Title, Course, Instructor Name and Duration.  Added the module name to the admin.py file to register it with the project.  From the web browser admin page, created three objects to be saved into a database utilizing the attributes of the djangoClasses attributes.  Added the classApp name to the main project settings.py file.  Made a finale Migration and Migrate command in the command prompt.  Ran the server to be able to modify and or delete the three objects in the admin page in the web browser.